Move CostReportConsumerMock to controller-api
